プラグインABIの実装ミスを修正し、SDKの不足機能を追加
ABI宣言と実装の不整合:
- Rustバインディングの CreateMemoryBuffer が解放関数の引数を欠いており、
  ホストがスタックのゴミを解放関数として保持し、GC時に不正なアドレスを
  呼び出していた問題を修正
- Rustバインディングの SetError が戻り値 bool を欠いていた問題を修正
- 誰も使用していない aosora_plugin_get_version の戻り値を void に統一
- PluginVersionInfo::flags が未初期化のままプラグインに渡っていた問題を修正。
  あわせてプラグイン側の出力値であることを明示するため pluginFlags に改名

SDKの不足機能:
- ValueWrapper::ToMemoryBuffer / to_memory_buffer を実装
- SetError / set_error をラッパから利用できるよう公開
- Rustの proc() がErrorオブジェクトをPluginErrorに作り変えていた問題を修正。
  Errorを継承したクラスも扱えるよう instance_of で判定する

その他:
- 引数の個数と添字を size_t から uint32_t に変更。size_tはメモリサイズ専用とする
- COMPATILBILITY_VERSION の綴りを COMPATIBILITY_VERSION に修正
- raw.rs の引数名をsnake_caseに統一

+++ aosora-shiori/CommonLibrary/PluginContextManager.cpp
@@ -355,11 +355,11 @@ namespace sakura {
 		}
 	}
 
-	size_t PluginContextManager::GetArgumentCount() {
-		return PeekContext().GetCallContext().args.size();
+	uint32_t PluginContextManager::GetArgumentCount() {
+		return static_cast<uint32_t>(PeekContext().GetCallContext().args.size());
 	}
 
-	aosora::raw::ValueHandle PluginContextManager::GetArgument(size_t index) {
+	aosora::raw::ValueHandle PluginContextManager::GetArgument(uint32_t index) {
 		if (index < GetArgumentCount()) {
 			return GetCurrentHandleManager().CreateHandle(
 				PeekContext().GetCallContext().args[index]
@@ -394,7 +394,7 @@ namespace sakura {
 		PeekContext().GetCallContext().threwError = ScriptValue::Make(pluginError);
 	}
 
-	void PluginContextManager::FunctionCall(aosora::raw::ValueHandle function, const aosora::raw::ValueHandle* argv, size_t argc) {
+	void PluginContextManager::FunctionCall(aosora::raw::ValueHandle function, const aosora::raw::ValueHandle* argv, uint32_t argc) {
 
 		//戻り値と例外を発生させる関数では既存の情報をクリア
 		PeekContext().GetCallContext();
@@ -407,7 +407,7 @@ namespace sakura {
 
 			//引数の展開
 			std::vector<ScriptValueRef> args;
-			for (size_t i = 0; i < argc; i++) {
+			for (uint32_t i = 0; i < argc; i++) {
 				args.push_back(GetCurrentHandleManager().GetValue(argv[i]));
 			}
 
@@ -424,7 +424,7 @@ namespace sakura {
 		}
 	}
 
-	aosora::raw::ValueHandle PluginContextManager::NewClassInstance(aosora::raw::ValueHandle classObject, const aosora::raw::ValueHandle* argv, size_t argc) {
+	aosora::raw::ValueHandle PluginContextManager::NewClassInstance(aosora::raw::ValueHandle classObject, const aosora::raw::ValueHandle* argv, uint32_t argc) {
 
 		//戻り値と例外を発生させる関数では既存の情報をクリア
 		PeekContext().GetCallContext();
@@ -435,7 +435,7 @@ namespace sakura {
 
 			//引数の展開
 			std::vector<ScriptValueRef> args;
-			for (size_t i = 0; i < argc; i++) {
+			for (uint32_t i = 0; i < argc; i++) {
 				args.push_back(GetCurrentHandleManager().GetValue(argv[i]));
 			}
